Move class definitions to separate file and simplfy mapping function
